+++ src/components/event-point.js
