+++ src/components/trip-info.js
