+++ src/main.js
@@ -1,11 +1,11 @@
-import {createMenuTemplate} from './view/site-menu';
-import {createFiltersTemplate} from './view/filters';
-import {createSortTemplate} from './view/sort-trip';
-import {createEventEditTemplate} from './view/event-edit';
-import {createEventPointTemplate} from './view/event-point';
-import {createEventTemplate} from './view/events-item';
-import {createTripInfoTemplate} from './view/trip-info';
-import {createTripCostTemplate} from './view/trip-cost';
+import {createMenuTemplate} from './components/site-menu';
+import {createFiltersTemplate} from './components/filters';
+import {createSortTemplate} from './components/sort-trip';
+import {createEventEditTemplate} from './components/event-edit';
+import {createEventPointTemplate} from './components/event-point';
+import {createEventTemplate} from './components/events-item';
+import {createTripInfoTemplate} from './components/trip-info';
+import {createTripCostTemplate} from './components/trip-cost';
 
 const EVENT_COUNT = 3;
